@@ -2,9 +2,9 @@
 Gem::Specification.new do |s|
   s.platform    = Gem::Platform::RUBY
   s.name        = 'spree_admin_roles_and_access'
-  s.version     = '3.1.0.beta1'
+  s.version     = '3.1.0'
   s.summary     = 'Dynamically defines roles and grants it permissions'
-  s.required_ruby_version = '>= 2.0.0'
+  s.required_ruby_version = '>= 2.1.0'
   s.files = Dir['LICENSE', 'README.md', 'app/**/*', 'config/**/*', 'lib/**/*', 'db/**/*']
 
   s.author    = "Nishant 'CyRo' Tuteja"
@@ -14,8 +14,8 @@ Gem::Specification.new do |s|
   s.require_path = 'lib'
   s.requirements << 'none'
 
-  s.add_dependency 'spree_core'
-  s.add_dependency 'spree_auth_devise'
+  s.add_dependency 'spree_core', '~> 3.1.0.beta'
+  s.add_dependency 'spree_auth_devise', '~> 3.1.0'
 
   s.add_development_dependency 'capybara'
   s.add_development_dependency 'ffaker'
