@@ -1,8 +1,8 @@
 module Spree
   module Admin
     class RolesController < ResourceController
-      before_filter :load_permissions, :only => [:edit, :new, :create, :update]
-      before_filter :restrict_unless_editable, :only => [:edit, :update]
+      before_filter :load_permissions, only: [:edit, :new, :create, :update]
+      before_filter :restrict_unless_editable, only: [:edit, :update]
 
       def index
         @roles = Spree::Role.page(params[:page])
@@ -10,7 +10,7 @@ module Spree
 
       private
         def permitted_resource_params
-          params.require(:role).permit(:name, :permission_ids => [])
+          params.require(:role).permit(:name, permission_ids: [])
         end
 
         def load_permissions
