@@ -4,11 +4,11 @@ module Spree
     def initialize(user)
       self.clear_aliased_actions
 
-      alias_action :edit, :to => :update
-      alias_action :new, :to => :create
-      alias_action :new_action, :to => :create
-      alias_action :show, :to => :read
-      alias_action :delete, :to => :destroy
+      alias_action :edit, to: :update
+      alias_action :new, to: :create
+      alias_action :new_action, to: :create
+      alias_action :show, to: :read
+      alias_action :delete, to: :destroy
 
       user ||= Spree.user_class.new
 
