@@ -1,6 +1,6 @@
 Spree::Role.class_eval do
 
-  has_and_belongs_to_many :permissions, :join_table => 'spree_roles_permissions', :class_name => 'Spree::Permission'
+  has_and_belongs_to_many :permissions, join_table: 'spree_roles_permissions', class_name: 'Spree::Permission'
 
   validates :name, uniqueness: true, allow_blank: true
 
@@ -14,5 +14,5 @@ Spree::Role.class_eval do
     permissions.pluck(:title).include?(permission_title)
   end
 
-  scope :default_role, lambda { where(:is_default => true) }
+  scope :default_role, lambda { where(is_default: true) }
 end
